Simplify the search box so that it works for users without Javascript enabled
The non javascript behaviour for the main GOV.UK search box is to inline the label and the search box, with the search box pushed to the right of the label.

This does not work for us, as our label ('Search the service manual') is considerably longer, and there is not space for the label and the input to co-exist on the same line.

This moves the label above the search box for all users.